# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RoleHeroes API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http:\\\\localhost:2609/api/",
            "description": "local server"
        },
        {
            "url": "https://test-role-heroes.herokuapp.com/api/",
            "description": "Test server"
        }
    ],
    "paths": {
        "/games": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Get all games",
                "operationId": "getGames",
                "responses": {
                    "200": {
                        "description": "All games"
                    }
                },
                "security": [
                    {
                        "scalar": "Bearer JWT"
                    }
                ]
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register user",
                "operationId": "authRegistry",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/RegisterUser"
                                    }
                                ]
                            },
                            "examples": {
                                "New User": {
                                    "summary": "Simple new user",
                                    "value": {
                                        "login": "new_user",
                                        "password": "qweqwe",
                                        "password_confirmation": "qweqwe"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success register",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/AccessToken"
                                        }
                                    ]
                                },
                                "examples": {
                                    "result": {
                                        "summary": "Succes response by register",
                                        "value": {
                                            "access_token": "example_token",
                                            "token_type": "brear",
                                            "expires_in": 604800
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "authLogin",
                "responses": {
                    "200": {
                        "description": "Success login"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AccessToken": {
                "title": "Access token info",
                "properties": {
                    "access_token": {
                        "title": "Access token",
                        "type": "string",
                        "format": "string"
                    },
                    "token_type": {
                        "title": "Token type",
                        "type": "string",
                        "format": "string"
                    },
                    "expires_in": {
                        "title": "Expire time",
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
            "RegisterUser": {
                "title": "New user",
                "properties": {
                    "login": {
                        "type": "string",
                        "example": "user_user"
                    },
                    "password": {
                        "type": "string",
                        "example": "qweqwe"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "qweqwe"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Auth by JWT",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Games",
            "description": "Methods for work with games"
        },
        {
            "name": "Auth",
            "description": "Methods auth for clients"
        }
    ]
}